Fix Neon migration: run per-statement, cast counts to int

+++ src/app/api/courses/[id]/access/route.ts
@@ -18,8 +18,8 @@ export async function GET(
 
     const courses = await sql`
       SELECT c.*,
-        (SELECT COUNT(*) FROM videos v WHERE v.course_id = c.id) as "videoCount",
-        (SELECT COUNT(*) FROM course_files f WHERE f.course_id = c.id) as "fileCount"
+        (SELECT COUNT(*)::int FROM videos v WHERE v.course_id = c.id) as "videoCount",
+        (SELECT COUNT(*)::int FROM course_files f WHERE f.course_id = c.id) as "fileCount"
       FROM courses c WHERE c.id = ${params.id}
     `;
     const course = courses[0];
